move userregs and syscallinfo to protocol::abi, related cleanups
this gives us a bit more infrastructure for detailed instruction tracing, with or without the usual trace logging.

+++ sand/src/ipc.rs
@@ -48,7 +48,7 @@ impl Socket {
     }
 
     extern "C" fn handle_sigio(num: u32) {
-        assert_eq!(num, abi::SIGIO);
+        assert_eq!(num, abi::SIGIO as u32);
         SIGIO_FLAG.store(true, Ordering::SeqCst);
     }
